Compatibility fix
The usage of UnityEditor.iOS.Xcode relies upon Unity's support for building iOS having been installed.

The postbuild process for iOS is also removed.

Example of what this solves: Multiple people checkout a branch which lacks these changes. This person mostly uses Unity in the editor, and has not installed the dependencies for building iOS and Android.

This pull request fixes the situation that would arise, wherein said person would be unable to work.